fix: close overwrite-guard symlink escapes and split run tests
Harden the --overwrite guard against symlink redirection: the target
must now resolve canonically where its spelling claims (realpath
compared against the normalized spelling, anchored at the project
directory), which refuses ancestor symlinks, symlink components hidden
behind '..' segments, and '..' escapes through a symlinked project
directory, while a project reached through a symlinked parent keeps
working. The alias refusal also compares canonical paths so symlink
spellings of the project directory or runs container cannot bypass it.

Make config_run_id honor its documented unreadable-config fallback by
catching OSError on the parse-only read, and classify the ecc check
run_dir display by canonical containment so symlink-escaping targets
display absolute while project-local '..foo' paths stay relative.

Split the run-directory/overwrite/alias tests out of test_run.py
(703 lines, over the review gate) into test_run_directory.py with the
shared run-execution mock as a commands/conftest.py fixture, and add
the plan-designated versioned spec docs/specification/run-directory.md.
